requirements.txt
pyspark==3.5.0
pyyaml==6.0.1
boto3==1.34.0
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.23
pyodbc==5.0.1
xxhash==3.4.1
tqdm==4.66.1
